requirement.txt
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.3
decorator==5.2.1
Django==5.2.6
djangorestframework==3.16.1
drf-spectacular==0.28.0
executing==2.2.1
idna==3.10
inflection==0.5.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyMySQL==1.1.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
